uart_tx: Go from idle to start state in 2 clock cycles
This is an improvement from before, where it took 2 baud cycles (20,832
clock cycles at 100 MHz and 9600 baud). This improvement was made by
turning off the baud counter when in the idle state.

When [i_start] is asserted while in the idle state, the uart_tx module
transitions into the start state and restarts the baud counter during
the first following clock cycle. During the second cycle, [o_busy] is
asserted and [o_tx] is deasserted to provide the start bit.